@@ -2,11 +2,13 @@
   "name": "paperie",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "clean": "rimraf .next"
   },
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
@@ -23,17 +25,21 @@
     "@radix-ui/react-tabs": "^1.1.11",
     "@radix-ui/react-tooltip": "^1.2.6",
     "axios": "^1.8.4",
+    "bcryptjs": "^3.0.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.5.2",
+    "input-otp": "^1.4.2",
     "iron-session": "^8.0.4",
     "lucide-react": "^0.484.0",
+    "mongodb": "^6.16.0",
+    "mongoose": "^8.15.0",
     "next": "^15.3.2",
     "qrcode": "^1.5.4",
-    "react": "^19.0.0",
+    "react": "^19.1.0",
     "react-day-picker": "^9.7.0",
-    "react-dom": "^19.0.0",
+    "react-dom": "^19.1.0",
     "react-hook-form": "^7.55.0",
     "sharp": "^0.34.1",
     "tailwind-merge": "^3.0.2",
@@ -46,6 +52,7 @@
     "compression-webpack-plugin": "^11.1.0",
     "eslint": "^9",
     "eslint-config-next": "15.2.4",
+    "rimraf": "^6.0.1",
     "tailwindcss": "^4"
   }
 }
