@@ -1,6 +1,7 @@
 import { invitationCards, envelopes } from "@/data/shop/products";
 import { notFound } from "next/navigation";
 import CheckoutForm from "@/components/ui/shop/checkout-form";
+import Image from "next/image";
 export default async function ProductPage({ params }) {
   const { slug } = await params;
 
@@ -14,10 +15,12 @@ export default async function ProductPage({ params }) {
     <div className="flex flex-col justify-center items-center  mx-auto p-4 pt-5 mt-2">
       <h1 className="text-4xl font-bold">{product.name}</h1>
       <div className="flex items-center ">
-        <img
+        <Image
+          width={300}
+          height={300}
           src={product.image}
           alt={product.name}
-          className="mt-3 ml-4 w-64 h-64 object-cover "
+          className="mt-3 ml-4 w-64 h-64 object-cover rounded-2xl "
         />
         <p className="ml-7 text-lg font-bold">{product.description}</p>
       </div>
