@@ -1,27 +1,32 @@
 import { GalleryVerticalEnd } from "lucide-react";
 
-import { LoginForm } from "@/components/login-form";
+import { SignUpForm } from "@/components/signup-form";
+import Link from "next/link";
+import Image from "next/image";
 
-export default function LoginPage() {
+const SignUpPage = () => {
   return (
     <div className="grid min-h-svh lg:grid-cols-2">
       <div className="flex flex-col gap-4 p-6 md:p-10">
         <div className="flex justify-center gap-2 md:justify-start">
-          <a href="#" className="flex items-center gap-2 font-medium">
+          <Link href="#" className="flex items-center gap-2 font-medium">
             <div className="bg-primary text-primary-foreground flex size-6 items-center justify-center rounded-md">
               <GalleryVerticalEnd className="size-4" />
             </div>
             Paperie Goa
-          </a>
+          </Link>
         </div>
         <div className="flex flex-1 items-center justify-center">
           <div className="w-full max-w-xs">
-            <LoginForm />
+            <SignUpForm />
           </div>
         </div>
       </div>
       <div className="bg-muted relative hidden lg:block">
-        <img
+        <Image
+          quality={100}
+          width={500}
+          height={500}
           src="/images/08cf49e6-0d9e-4035-8887-7ef44eac7d71.jpeg"
           alt="Image"
           className="absolute inset-0 h-full w-full object-cover dark:brightness-[0.2] dark:grayscale"
@@ -29,4 +34,5 @@ export default function LoginPage() {
       </div>
     </div>
   );
-}
+};
+export default SignUpPage;
