@@ -20,7 +20,7 @@ const Testimonials = () => {
   return (
     <div className="px-4 py-8">
       <h1 className=" text-4xl font-semibold sm:text-4xl md:text-5xl text-center mb-4">
-        Testimonials from Our Clients
+        Hear from what some of Our Clients say about Us
       </h1>
       <p className=" transition-discrete text-4xl sm:text-5xl md:text-3xl lg:text-7xl px-4 pb-4 text-center">
         &quot;{quotes[currentIndex]}&quot;
